pyproject.toml
[tool.poetry]
name = "imagine"
version = "0.1.0"
description = ""
authors = ["dnl-strnl <129630837+dnl-strnl@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "imagine", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
hydra-core = "^1.3.2"
flask = "^3.1.0"
pillow = "^11.1.0"
omegaconf = "^2.3.0"
numpy = "^2.2.1"
requests = "^2.32.3"
diffusers = "^0.35.1"
torch = "^2.8.0"
accelerate = "^1.10.0"
transformers = "^4.55.3"
litserve = "^0.2.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
